requirements.txt
fastapi
uvicorn
pydantic
numpy
opencv-python-headless
tensorflow-cpu
sympy
python-multipart